Bug fixes and one addition
Fixed bug with random array indexing always missing the last element
of the array.

Fixed the wording of one of the commandments.

Added in a !reset command, which will hopefully let us fix his timers
without having to shut him all the way down, until we can find the bug
and fix it the right way.