Added '$badword' command
Added a command to add a new bad word to Timmy's database, to keep
his Markhov Chain data from getting offensive words that pollute
it. When a new word is added, it also scrubs it from the existing
Markhov Chains.

TODO: Maybe use this badword list for the chain novel and get
command input as well.